cleanup: Consolidate to a single `delay` method.
