+++ src/main/utils/delay.ts
@@ -1,5 +1,7 @@
-function delay(timeInMs: number) {
+/**
+ * Waits the specified amount of milliseconds
+ * @param {number} timeInMs - The amount of time to wait in milliseconds
+ */
+export function delay(timeInMs: number) {
   return new Promise(resolve => setTimeout(resolve, timeInMs));
 }
-
-export default delay;
